+++ index.d.ts
@@ -1,3 +1,4 @@
 import { Options } from './types';
 export declare function getAffectedSync(patternArg: string | Options, optionsArg?: Options): string[];
-export default getAffectedSync;
+export declare function getAffected(patternArg: string | Options, optionsArg?: Options): Promise<string[]>;
+export default getAffected;