+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.0
+
+- breaking: default exported function is async now
+- getAffectedSync moved to named export
+
 ## 2.10.0
 
 - rename `options.superleaves` to `options.usink`, mark `options.superleaves` as deprecated